+++ test/test_write.py
@@ -29,8 +29,9 @@ def test_basic_write(
 
 
 def test_basic_write_no_file(tmp_path: Path) -> None:
-    basic_write(project_dir=tmp_path, version="1.2.3", params={})
-    assert list(tmp_path.iterdir()) == []
+    with pytest.raises(ConfigError) as excinfo:
+        basic_write(project_dir=tmp_path, version="1.2.3", params={})
+    assert str(excinfo.value) == "tool.versioningit.write.file must be set to a string"
 
 
 def test_basic_write_bad_ext(tmp_path: Path) -> None: